extracted zoom behavior to its own class
